+++ packages/gamut/src/Dialog/index.tsx
@@ -1,24 +1,35 @@
 import { MiniDeleteIcon } from '@codecademy/gamut-icons';
 import { system } from '@codecademy/gamut-styles';
-import { useTheme } from '@emotion/react';
+import { StyleProps } from '@codecademy/variance';
 import styled from '@emotion/styled';
 import React from 'react';
 
-import { FlexBox } from '../Box';
 import { FillButton, IconButton, TextButton } from '../Button';
 import { Overlay, OverlayProps } from '../Overlay';
 import { Pattern } from '../Pattern';
 import { Text } from '../Typography';
 
+type SizeVariants = StyleProps<typeof modalSizeVariants>;
+
+const modalSizeVariants = system.variant({
+  prop: 'size',
+  variants: {
+    small: { width: '400px', minHeight: '170px' },
+    medium: { width: '640px', minHeight: '240px' },
+  },
+});
+
 const ShroudedOverlay = styled(Overlay)(system.css({ bg: 'shadow' }));
 
-const ModalWrapper = styled.div(
+const ModalWrapper = styled.div<SizeVariants>(
   system.css({
     position: 'relative',
     zIndex: 1,
     maxWidth: 'calc(100vw - 2rem)',
     borderRadius: '2px',
-  })
+    display: 'grid',
+  }),
+  modalSizeVariants
 );
 
 /** This will need to be consolidated with Card / CoachMark / Toast  */
@@ -44,31 +55,16 @@ const ModalBody = styled.div(
     rowGap: 12,
     columnGap: 16,
     gridTemplateColumns: '1fr min-content 2rem',
-    gridTemplateRows: 'repeact(3, auto)',
+    gridTemplateRows: 'max-content 1fr max-content',
     gridTemplateAreas: `'title title close'
-    'children children children'
+    'content content content'
     'cancel confirm confirm'`,
   })
 );
 
-const CloseButton = styled.div`
-  grid-area: close;
-`.withComponent(IconButton);
-
-const ConfirmButton = styled(FillButton)`
-  grid-area: confirm;
-`;
-
-const CancelButton = styled(TextButton)`
-  grid-area: cancel;
-`.withComponent((props: DialogProps['cancelCta']) => (
-  <FlexBox justifyContent="flex-end">
-    <TextButton {...props} />
-  </FlexBox>
-));
-
 export interface DialogProps
-  extends Pick<OverlayProps, 'clickOutsideCloses' | 'escapeCloses'> {
+  extends Pick<OverlayProps, 'clickOutsideCloses' | 'escapeCloses'>,
+    SizeVariants {
   isOpen: boolean;
   title: React.ReactNode;
   children: React.ReactNode;
@@ -89,12 +85,9 @@ export const Dialog: React.FC<DialogProps> = ({
   confirmCta,
   cancelCta,
   onRequestClose,
+  size = 'small',
   ...rest
 }) => {
-  const {
-    colorModes: { active },
-  } = useTheme();
-
   const onConfirm = () => {
     onRequestClose();
     confirmCta.onClick?.();
@@ -107,26 +100,37 @@ export const Dialog: React.FC<DialogProps> = ({
 
   return (
     <ShroudedOverlay onRequestClose={onCancel} {...rest}>
-      <ModalWrapper aria-hidden="false" aria-modal="true" role="dialog">
+      <ModalWrapper
+        size={size}
+        aria-hidden="false"
+        aria-modal="true"
+        role="dialog"
+      >
         <ModalShadow name="checkerDense" />
         <ModalBody>
           <Text as="h2" fontSize={20} lineHeight="base" gridArea="title">
             {title}
           </Text>
-          <CloseButton
+          <IconButton
             aria-label="Close Dialog"
-            mode={active}
             size="small"
+            variant="secondary"
             icon={MiniDeleteIcon}
             onClick={onCancel}
+            gridArea="close"
           />
-          <Text as="p" gridArea="children">
+          <Text as="p" gridArea="content">
             {children}
           </Text>
           {cancelCta && (
-            <CancelButton mode={active} {...cancelCta} onClick={onCancel} />
+            <TextButton
+              {...cancelCta}
+              onClick={onCancel}
+              justifySelf="end"
+              gridArea="cancel"
+            />
           )}
-          <ConfirmButton mode={active} {...confirmCta} onClick={onConfirm} />
+          <FillButton {...confirmCta} onClick={onConfirm} gridArea="confirm" />
         </ModalBody>
       </ModalWrapper>
     </ShroudedOverlay>