+++ packages/gamut/src/Form/Select.tsx
@@ -67,7 +67,7 @@ const allowClickStyle = css`
   pointer-events: none;
 `;
 
-export const StyledFlexbox = styled(FlexBox)(allowClickStyle);
+const StyledFlexbox = styled(FlexBox)(allowClickStyle);
 
 export const Select = forwardRef<HTMLSelectElement, SelectWrapperProps>(
   (
@@ -94,12 +94,13 @@ export const Select = forwardRef<HTMLSelectElement, SelectWrapperProps>(
         className={className}
       >
         <StyledFlexbox
-          paddingRight={12}
+          pr={12}
           alignItems="center"
           position="absolute"
           right="0"
           top="0"
           bottom="0"
+          aria-hidden
         >
           {sizeVariant === 'small' ? (
             <MiniChevronDownIcon size={12} />